Fix issue in pace timeline with no moving time (#23)
Sometimes the "moving time" of an activity is broken in Strava.
Typically for very old Garmin uploads, this field is either zero or way
too large. If it is zero, when we resample to create the moving average
lines, we can divide by zero, and get extra NaN's and then filter them
out. Very niche issue, but this takes care of it + adds a test to make
sure we don't regress.