Redirect to download-strava-data error
Whenever I redirected from /home to /home/download_strava_data, I got
a whopping great big error that was stopping me from being able to
fix other errors because I didn't know which was which. Using the router
in the set state, and not the rendering of the component seems to have
fixed it though.