Handle missing cookies issue
There is an issue somewhere that results in users requesting data from
this website without session cookies. I could reproduce on my phone by
logging in, hard-closing the browser, and then re-opening it. On re-open
the browser attempts to load the 'logged in' screen with no cookies.

This PR does the following:

1: Delete a whole lot of random stuff from the backend. We don't need
any of the html, css, and javascript any more since next.js handles
everything now.
2: Wraps all fetch requests so that if a 401 Unauthorized is received,
then the user gets re-routed to home.
3: The backend routes now have a decorator to check for the session
cookie instead of doing it in each route individually, and instead of
manually rerouting to 'index' which doesn't exist, they just return
401 status codes.