Fix polyline plot errors (#36)
There are two errors for polylines that have popped up on Sentry lately.
It seems that people have activities with either a single point, or
multiple points directly on top of eachother, which results in div/0
errors. This fixes those, and adds tests for the polyline visualisations
while I'm at it.