General Refactor (#39)
* Route Creation Centralized

Now the route creation is centralised in Tab, and just inheritted by
all subtabs, so all subtabs just need to focus on creating their JSON
responses, and not the intricacies of Flask route creation. Less copied
code.

* Util to flatten the gui_tree structure

Remove some duplication by providing an easy way to get a flattened
version of the gui_tree

* Frontend tab error looks better

Max page height is now 100% and not 100% + 1/3*100% margin.

* Actually fix the error screen

* Move tab logic into tabs and out of server / queue

Big refactor move the logic around 'what should we do for each tab re-
processing, and viewing. Now instead of an expanding list of if/else
statements in the task queue, there is one if/else in each tab! Yay!

* Fix github icon alignment in firefox/safari

* Shuffle tab files around + fix tidbit local bug

* Mock s3 and add tests ensuring data is entering s3

This allows us to check that plot, table, and image data is serialised
correctly allowing it to be placed into S3.

* Generalise Cumulative Plots

I generalised cumulative plots. So instead of having two plot files for
cumulative time and cumulative distance, now I have 1 file for
cumulative time, distance, elevation, and kudos.

* Fix test using wrong function