Attempt to add an image to the metadata of the home page (#45)
When sending this site via fb messenger or linking / sharing on other
websites, this image should be displayed. May or may not work because
apparently it only works as a server side component, but I'm exporting,
so I genuinely have no idea if this will work, but let's just send it.